update LoL data update tick

+++ Project-Aurora/Project-Aurora/Profiles/LeagueOfLegends/GSI/GameData.cs
@@ -66,14 +66,14 @@ namespace AuroraRgb.Profiles.LeagueOfLegends.GSI
     public class _Rune
     {
         public string displayName { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string rawDescription { get; set; }
         public string rawDisplayName { get; set; }
     }
 
     public class _StatRune
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string rawDescription { get; set; }
     }
